+++ Clase Springboot/src/main/java/com/uade/tpo/demo/service/FuncionService.java
@@ -1,5 +1,5 @@
 package com.uade.tpo.demo.service;
-
+import java.math.BigDecimal;
 import com.uade.tpo.demo.entity.Funcion;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface FuncionService {
     List<Funcion> getFunciones();
     Optional<Funcion> getFuncionById(Long id);
-    Funcion createFuncion(Long peliculaId, Long salaId, LocalDateTime horario, String idioma, String formato, Float precioBase);
+    Funcion createFuncion(Long peliculaId, Long salaId, LocalDateTime horario, String idioma, String formato, BigDecimal precioBase);
 }
